requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
loguru
pytesseract
pillow
opencv-python